requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
numpy==2.1.2
Pillow==10.4.0
tqdm==4.66.5
open_clip_torch==2.26.1
faiss-cpu==1.9.0
